feat: extract address field into separate component
