+++ src/components/app/updateUserProfileForm/step1/constants.ts
@@ -1,12 +1,9 @@
-import { ORDERED_SUPPORTED_COUNTRIES } from '@/utils/shared/supportedCountries'
+import { SupportedCountryCodes } from '@/utils/shared/supportedCountries'
 
 export const DEFAULT_DISCLAIMER =
   'You are about to change your viewing country based on the address you have chosen'
 
-export const DISCLAIMERS_BY_COUNTRY_CODE: Record<
-  (typeof ORDERED_SUPPORTED_COUNTRIES)[number],
-  string
-> = {
+export const DISCLAIMERS_BY_COUNTRY_CODE: Record<SupportedCountryCodes, string> = {
   us: 'You are about to change your viewing country based on the address you have chosen',
   gb: 'You are about to change your viewing country based on the address you have chosen',
   ca: 'You are about to change your viewing country based on the address you have chosen',
